MISUV-8129 Move OptOutContextData to OptOutSessionDataRepository for better encapsulation
