+++ app/models/optout/OptOutSessionData.scala
@@ -17,6 +17,7 @@
 package models.optout
 
 import play.api.libs.json.{Json, OFormat}
+import repositories.OptOutContextData
 
 case class OptOutSessionData(optOutContextData: Option[OptOutContextData],
                              selectedOptOutYear: Option[String]) {
